Rework LogMonitor
