+++ field_friend/interface/components/development.py
@@ -7,7 +7,7 @@ from nicegui import ui
 
 from .hardware_control import create_hardware_control_ui
 from .io_overview import IoOverview as io_overview
-from .log_view import LogView
+from .log_monitor import LogMonitor
 from .settings import create_settings_ui
 from .status_dev import status_dev_page
 
@@ -32,7 +32,6 @@ def create_development_ui(system: System) -> None:
                         rosys.simulation_ui()
                 create_hardware_control_ui(system.field_friend, system.automator, system.puncher)
                 status_dev_page(system.field_friend, system)
-                LogView().classes('w-100 h-60')
 
     with ui.row():
         with ui.card():
@@ -49,4 +48,6 @@ def create_development_ui(system: System) -> None:
                 esp_pins_p0 = system.field_friend.robot_brain.esp_pins_p0
                 esp_pins_p0.developer_ui()
 
+    with ui.row().classes('w-1/2'):
+        LogMonitor().ui()
     io_overview(system)
